requirements.txt
asgiref
Django
django-filter
djangorestframework
drf-yasg
inflection
packaging
python-dotenv
pytz
PyYAML
sqlparse
tzdata
uritemplate

# Base de datos PostgreSQL
psycopg2-binary
PyMySQL
# Para despliegue
dj-database-url
gunicorn
whitenoise

